cap_api/requirements.txt
requests==2.25.1
flask==1.1.2
uniswap-python==0.4.4